Extract DirectPaymentsHelper in BankSystem.Web
